Update lcd_api.py
One linefeed too many, in case the line is filled up to the last character

In case I fill the line to the last row, it skips the following line without replacement and the following line remains empty. This is independent of whether I put a \n at the end of the line or not. If I fill the display only with 19 characters instead of the 20 possible characters, the line jump to the following line works.

I walked through your code and the following change in lcd_api.py in "putchar" definition will fix this issue: